Create Automate-Inventory.ps1
This PowerShell script converts all .xlsx files in the Working folder and converts it into .csv for processing. From there, it looks at the main category for Asset Types to create a .csv file for each Asset Type so that they can be uploaded into a database that only accepted one Asset Type at a time.